+++ src/Ext/House/Body.h
@@ -170,5 +170,5 @@ public:
 	static CanBuildResult BuildLimitGroupCheck(const HouseClass* pThis, const TechnoTypeClass* pItem, bool buildLimitOnly, bool includeQueued);
 	static bool ReachedBuildLimit(const HouseClass* pHouse, const TechnoTypeClass* pType, bool ignoreQueued);
 
-	static bool CanSelectOwner(HouseClass* pThis, HouseClass* pPlayer);
+	static bool CanBeSelectedBy(HouseClass* pOwner, HouseClass* pSelectingPlayer);
 };